Rename ActionWrapper to ActionMain
