+++ plugin-skel/root-saga.js
@@ -2,7 +2,7 @@ import { takeLatest, takeEvery } from "redux-saga";
 import { call, put, select } from "redux-saga/effects";
 
 import defaultSaga from "plugins-saga";
-import * as actions from "actions/ActionWrapper";
+import * as actions from "actions/ActionMain";
 import { postRequest, checkStatus, putCsrfToken } from "sagas/common";
 
 // define sagas
